+++ transport/grpc/config.go
@@ -66,15 +66,19 @@ func TransportSpec(opts ...Option) yarpcconfig.TransportSpec {
 //        exponential:
 //          first: 10ms
 //          max: 30s
+//      clientMaxHeaderListSize: 1024
+//      serverMaxHeaderListSize: 2048
 //
 // All parameters of TransportConfig are optional. This section
 // may be omitted in the transports section.
 type TransportConfig struct {
-	ServerMaxRecvMsgSize int                 `config:"serverMaxRecvMsgSize"`
-	ServerMaxSendMsgSize int                 `config:"serverMaxSendMsgSize"`
-	ClientMaxRecvMsgSize int                 `config:"clientMaxRecvMsgSize"`
-	ClientMaxSendMsgSize int                 `config:"clientMaxSendMsgSize"`
-	Backoff              yarpcconfig.Backoff `config:"backoff"`
+	ServerMaxRecvMsgSize    int                 `config:"serverMaxRecvMsgSize"`
+	ServerMaxSendMsgSize    int                 `config:"serverMaxSendMsgSize"`
+	ServerMaxHeaderListSize uint32              `config:"serverMaxHeaderListSize"`
+	ClientMaxRecvMsgSize    int                 `config:"clientMaxRecvMsgSize"`
+	ClientMaxSendMsgSize    int                 `config:"clientMaxSendMsgSize"`
+	ClientMaxHeaderListSize uint32              `config:"clientMaxHeaderListSize"`
+	Backoff                 yarpcconfig.Backoff `config:"backoff"`
 }
 
 // InboundConfig configures a gRPC Inbound.
@@ -279,6 +283,12 @@ func (t *transportSpec) buildTransport(transportConfig *TransportConfig, _ *yarp
 	}
 	if transportConfig.ClientMaxSendMsgSize > 0 {
 		options = append(options, ClientMaxSendMsgSize(transportConfig.ClientMaxSendMsgSize))
+	}
+	if transportConfig.ServerMaxHeaderListSize > 0 {
+		options = append(options, ServerMaxHeaderListSize(transportConfig.ServerMaxHeaderListSize))
+	}
+	if transportConfig.ClientMaxHeaderListSize > 0 {
+		options = append(options, ClientMaxHeaderListSize(transportConfig.ClientMaxHeaderListSize))
 	}
 	backoffStrategy, err := transportConfig.Backoff.Strategy()
 	if err != nil {