+++ transport/grpc/options.go
@@ -39,10 +39,11 @@ const (
 	// defensive programming
 	// these are copied from grpc-go but we set them explicitly here
 	// in case these change in grpc-go so that yarpc stays consistent
-	defaultServerMaxRecvMsgSize = 1024 * 1024 * 4
-	defaultServerMaxSendMsgSize = math.MaxInt32
-	defaultClientMaxRecvMsgSize = 1024 * 1024 * 4
-	defaultClientMaxSendMsgSize = math.MaxInt32
+	defaultServerMaxRecvMsgSize    = 1024 * 1024 * 4
+	defaultServerMaxSendMsgSize    = math.MaxInt32
+	defaultClientMaxRecvMsgSize    = 1024 * 1024 * 4
+	defaultClientMaxSendMsgSize    = math.MaxInt32
+	defaultServerMaxHeaderListSize = 1024 * 1024 * 16
 )
 
 // Option is an interface shared by TransportOption, InboundOption, and OutboundOption
@@ -108,6 +109,16 @@ func ServerMaxSendMsgSize(serverMaxSendMsgSize int) TransportOption {
 	}
 }
 
+// ServerMaxHeaderListSize returns a transport option for configuring maximum
+// header list size the server must accept.
+//
+// The default is 16MB (gRPC default).
+func ServerMaxHeaderListSize(serverMaxHeaderListSize uint32) TransportOption {
+	return func(transportOptions *transportOptions) {
+		transportOptions.serverMaxHeaderListSize = &serverMaxHeaderListSize
+	}
+}
+
 // ClientMaxRecvMsgSize is the maximum message size the client can receive.
 //
 // The default is 4MB.
@@ -123,6 +134,16 @@ func ClientMaxRecvMsgSize(clientMaxRecvMsgSize int) TransportOption {
 func ClientMaxSendMsgSize(clientMaxSendMsgSize int) TransportOption {
 	return func(transportOptions *transportOptions) {
 		transportOptions.clientMaxSendMsgSize = clientMaxSendMsgSize
+	}
+}
+
+// ClientMaxHeaderListSize returns a transport option for configuring maximum
+// header list size the client must accept.
+//
+// The default is 16MB (gRPC default).
+func ClientMaxHeaderListSize(clientMaxHeaderListSize uint32) TransportOption {
+	return func(transportOptions *transportOptions) {
+		transportOptions.clientMaxHeaderListSize = &clientMaxHeaderListSize
 	}
 }
 
@@ -190,13 +211,15 @@ func KeepaliveParams(params keepalive.ClientParameters) DialOption {
 }
 
 type transportOptions struct {
-	backoffStrategy      backoff.Strategy
-	tracer               opentracing.Tracer
-	logger               *zap.Logger
-	serverMaxRecvMsgSize int
-	serverMaxSendMsgSize int
-	clientMaxRecvMsgSize int
-	clientMaxSendMsgSize int
+	backoffStrategy         backoff.Strategy
+	tracer                  opentracing.Tracer
+	logger                  *zap.Logger
+	serverMaxRecvMsgSize    int
+	serverMaxSendMsgSize    int
+	clientMaxRecvMsgSize    int
+	clientMaxSendMsgSize    int
+	serverMaxHeaderListSize *uint32
+	clientMaxHeaderListSize *uint32
 }
 
 func newTransportOptions(options []TransportOption) *transportOptions {