+++ CHANGELOG.md
@@ -4,8 +4,7 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
 
-## [Unreleased]
-- No changes yet.
+## [1.61.0] - 2022-07-19
 ### Added
 - grpc - option to set custom maximum header list size.
 
@@ -1418,7 +1417,7 @@ This release requires regeneration of ThriftRW code.
 
 - Initial release.
 
-[Unreleased]: https://github.com/yarpc/yarpc-go/compare/v1.60.0...HEAD
+[1.61.0]: https://github.com/yarpc/yarpc-go/compare/v1.60.0...v1.61.0
 [1.60.0]: https://github.com/yarpc/yarpc-go/compare/v1.59.0...v1.60.0
 [1.59.0]: https://github.com/yarpc/yarpc-go/compare/v1.58.0...v1.59.0
 [1.58.0]: https://github.com/yarpc/yarpc-go/compare/v1.57.1...v1.58.0